requirements.txt
streamlit
openai
faiss-cpu
pdfplumber
tiktoken
pandas
